Version 1.1 con manejo de errores y organización en el codigo, adémas ahora se manejan numeros grandes en las conversiones.

+++ src/main/java/src/Moneda.java
@@ -3,6 +3,6 @@ package src;
 public record Moneda(String result,
                      String base_code,
                      String target_code,
-                     float conversion_rate,
-                     float conversion_result) {
+                     double conversion_rate,
+                     double conversion_result) {
 }
